test(table group): table group 인수 테스트
