+++ src/test/java/kitchenpos/acceptance/tablegroup/TableGroupAcceptance.java
@@ -1,9 +1,8 @@
-package kitchenpos.acceptance.table;
+package kitchenpos.acceptance.tablegroup;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import org.assertj.core.api.AbstractIntegerAssert;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
@@ -11,22 +10,23 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.acceptance.util.AcceptanceTest;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
 
 public class TableGroupAcceptance extends AcceptanceTest {
 
 	public static final String TABLE_GROUP_REQUEST_URL = "/api/table-groups";
 
-	public static ExtractableResponse<Response> 단체_지정_등록_요청(TableGroup tableGroup) {
+	public static ExtractableResponse<Response> 단체_지정_등록_요청(TableGroupRequest request) {
 		return RestAssured
 			.given().log().all()
-			.body(tableGroup)
+			.body(request)
 			.contentType(MediaType.APPLICATION_JSON_VALUE)
 			.when().post(TABLE_GROUP_REQUEST_URL)
 			.then().log().all().extract();
 	}
 
-	public static ExtractableResponse<Response> 단체_지정_해제_요청(TableGroup expected) {
+	public static ExtractableResponse<Response> 단체_지정_해제_요청(TableGroupResponse expected) {
 		return RestAssured
 			.given().log().all()
 			.when().delete(TABLE_GROUP_REQUEST_URL + "/{tableGroupId}", expected.getId())
@@ -36,13 +36,13 @@ public class TableGroupAcceptance extends AcceptanceTest {
 	public static void 단체_지정_등록됨(ExtractableResponse<Response> response) {
 		assertAll(
 			() -> assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value()),
-			() -> assertThat(response.as(TableGroup.class)).isNotNull(),
-			() -> assertThat(response.as(TableGroup.class).getId()).isNotNull(),
-			() -> assertThat(response.as(TableGroup.class).getCreatedDate()).isNotNull()
+			() -> assertThat(response.as(TableGroupResponse.class)).isNotNull(),
+			() -> assertThat(response.as(TableGroupResponse.class).getId()).isNotNull(),
+			() -> assertThat(response.as(TableGroupResponse.class).getCreatedDate()).isNotNull()
 		);
 	}
 
-	public static AbstractIntegerAssert<?> 단체_지정_해제됨(ExtractableResponse<Response> response) {
-		return assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+	public static void 단체_지정_해제됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
 	}
 }
